backend/requirements.txt
fastapi==0.115.0
fastapi-mail==1.4.1
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
httpx==0.27.2
python-multipart==0.0.12
